tool_relateer/news_relations/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.0.267
langchain-community==0.0.13
anthropic==0.5.0
openai==1.3.0
beautifulsoup4==4.12.2
lxml==4.9.3
plotly==5.17.0
networkx==3.1
pyvis==0.3.2 